# docs/swagger.yaml
openapi: 3.0.2
info:
    title: BookEST
    version: 1.0.0
paths:
    /rooms:
        get:
            tags:
                - rooms
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                list of rooms:
                                    value:
                                        - Suites
                                        - Double
                                        - Twin
                    description: return of list of rooms
            summary: get a list of rooms from api
            description: returns a list of all rooms in the API
        post:
            requestBody:
                description: a new room to be created
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/new-room'
                        examples:
                            example of a room:
                                value:
                                    name: some text
                                    price: 8.51
                                    description: some text
                required: true
            tags:
                - new-room
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new-room'
                    description: Rooms has been created into the array/database
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                One or all details are missing:
                                    value:
                                        error: One or all details are missing
                    description: NOt found
            description: one or more not found
    '/rooms/{id}':
        get:
            tags:
                - details
            parameters:
                -
                    examples:
                        Id of a room:
                            value: '48'
                    name: id
                    description: Id of the room
                    schema:
                        type: integer
                    in: path
                    required: true
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Details of a room:
                                    value: |-
                                        [
                                           "id": 1,
                                               "name": "double",
                                                "price": 111.50,
                                                "decription": "with view"
                                        ]
                    description: 'Returns full details of a given room, as determined by the id'
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                room not found:
                                    value:
                                        error: room not found wit given id
                    description: Room not found with given Id
            operationId: id
            summary: Get full details of a room
            description: Returns full detauils about a specific room as determined by id input put the user
        delete:
            tags:
                - delete-room
            responses:
                '204':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/delete-room'
                            examples:
                                delete:
                                    value:
                                        id: 5
                    description: 'Deleted successfully, nothing to display'
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                error:
                                    value:
                                        error: 'Can''t delete, id not found'
                    description: room not found
            summary: Delete rooms with given Id
            description: Delete rooms with given Id
        parameters:
            -
                examples:
                    integer of requested id:
                        value: '63'
                name: id
                description: Id of the room that we require
                schema:
                    type: integer
                in: path
                required: true
components:
    schemas:
        new-room:
            title: Root Type for new-room
            description: ''
            required:
                - price
                - name
                - description
            type: object
            properties:
                name:
                    type: string
                price:
                    format: float
                    type: number
                description:
                    type: string
            example:
                name: Presidental suite
                price: 99999
                description: Extra king sized bed with bath view
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message
        delete-room:
            description: Delete room with given id
            required:
                - id
            type: object
            properties:
                id:
                    description: id of the room
                    type: integer
            example:
                id: 34